--- resources/js/Components/providers/AppContextProvider/index.tsx
@@ -1,3 +1,0 @@
-import AppContextProvider from "./AppContextProvider";
-
-export default AppContextProvider;
